// types.ts
export interface Reward {
  id: number;
  title: string;
  partner: string;
  cost: number;
  imageUrl: string;
  rewardType?: 'digital_wallet';
  priceNPR?: number;
  listingTier: 'Gold' | 'Silver' | 'Bronze';
}